app/src/main/java/com/example/concordia_campus_guide/models/Time.java
package com.example.concordia_campus_guide.models;

import java.io.Serializable;

public class Time implements Serializable {
    private String text;
    private String timeZone;
    private long value;

    public String getText() {
        return text;
    }

    public void setText(String text) {
        this.text = text;
    }

    public String getTimeZone() {
        return timeZone;
    }

    public void setTimeZone(String timeZone) {
        this.timeZone = timeZone;
    }

    public long getValue() {
        return value;
    }

    public void setValue(long value) {
        this.value = value;
    }
}
